pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "health-rag"
version = "0.1.0"
description = "Local-first Telegram mini-RAG for health documents"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=21.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
ui = [
    "gradio>=4.44.0",
]

[project.scripts]
health-rag-ingest = "health_rag.cli.ingest:main"
health-rag-bot = "health_rag.telegram_app.bot:main"
health-rag-ui = "health_rag.ui.gradio_app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
